+++ cli/src/pcluster/cli.py
@@ -82,6 +82,14 @@ def build_image(args):
     pcluster.build_image(args)
 
 
+def delete_image(args):
+    pcluster.delete_image(args)
+
+
+def describe_image(args):
+    pcluster.describe_image(args)
+
+
 def config_logger():
     logger = logging.getLogger("pcluster")
     file_only_logger = logging.getLogger("cli_log_file")
@@ -307,6 +315,28 @@ Returns an ssh command with the cluster username and IP address pre-populated::
     _addarg_config(pami)
     _addarg_region(pami)
     pami.set_defaults(func=build_image)
+
+    # delete image command subparser
+    pdeleteami = subparsers.add_parser(
+        "delete-image",
+        help="Deletes an image and related image builder stack.",
+        epilog="When the command is called and it begins polling for the status of that call "
+        'it is safe to "Ctrl-C" out.',
+    )
+    pdeleteami.add_argument(
+        "-i", "--image-name", dest="image_name", required=True, help="Name of the AWS ParallelCluster AMI to delete."
+    )
+    _addarg_region(pdeleteami)
+    _addarg_nowait(pdeleteami)
+    pdeleteami.set_defaults(func=delete_image)
+
+    # describe image command subparser
+    pdescribeami = subparsers.add_parser("describe-image", help="Describes the specified ParallelCluster image.")
+    pdescribeami.add_argument(
+        "-i", "--image-name", dest="image_name", required=True, help="Name of the AWS ParallelCluster AMI to describe."
+    )
+    _addarg_region(pdescribeami)
+    pdescribeami.set_defaults(func=describe_image)
 
     # configure command subparser
     pconfigure = subparsers.add_parser("configure", help="Start the AWS ParallelCluster configuration.")